Print an EOL before printing the TestResult while executing the spec
By printing this newline, the filepath of the failing assertion does not get
preceded by the dot which represents a running test.

This gives the advantage, that the complete output of unity can be used as it is
with a makefile in vim. Every error gets displayed in the quickfix and you can
jump appropiately.